added header

+++ hwpe-tb/sw/aes_hwpe.c
@@ -12,7 +12,7 @@ void aes_hwpe_configure(uint8_t *input, uint8_t *output, uint8_t key_length)
 
     return;
 }
-void aes_hwpe_key_set(uint8_t *key_value)
+void aes_hwpe_key_set(uint32_t *key_value)
 {
     hwpe_key_set(key_value);
 }
